requirements.txt
guardrails-ai
openai
IPython
presidio-analyzer 
presidio-anonymizer